+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/events.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "common.h"
-
 #include <ydb/core/formats/arrow/arrow_filter.h>
 #include <ydb/core/tx/columnshard/columnshard_private_events.h>
 #include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/source.h>
@@ -33,46 +31,6 @@ private:
 
 public:
     TEvRequestFilter(const IDataSource& source, const std::shared_ptr<IFilterSubscriber>& subscriber);
-};
-
-class TEvFilterConstructionResult
-    : public NActors::TEventLocal<TEvFilterConstructionResult, NColumnShard::TEvPrivate::EvFilterConstructionResult> {
-private:
-    using TFilters = THashMap<TDuplicateMapInfo, NArrow::TColumnFilter>;
-    TConclusion<TFilters> Result;
-
-public:
-    TEvFilterConstructionResult(TConclusion<TFilters>&& result)
-        : Result(std::move(result)) {
-        if (Result.IsSuccess()) {
-            for (const auto& [info, filter] : *Result) {
-                AFL_VERIFY(!!filter.GetRecordsCount() && filter.GetRecordsCountVerified() == info.GetRowsCount())(
-                                                                                             "filter", filter.GetRecordsCount().value_or(0))(
-                                                                                             "info", info.DebugString());
-            }
-        }
-    }
-
-    const TConclusion<TFilters>& GetConclusion() const {
-        return Result;
-    }
-
-    TFilters&& ExtractResult() {
-        return Result.DetachResult();
-    }
-};
-
-class TEvDuplicateFilterDataFetched
-    : public NActors::TEventLocal<TEvDuplicateFilterDataFetched, NColumnShard::TEvPrivate::EvDuplicateFilterDataFetched> {
-private:
-    YDB_READONLY_DEF(ui64, SourceId);
-    YDB_READONLY(TConclusion<TColumnsData>, Result, TConclusionStatus::Success());
-
-public:
-    TEvDuplicateFilterDataFetched(const ui64 sourceId, TConclusion<TColumnsData>&& result)
-        : SourceId(sourceId)
-        , Result(std::move(result)) {
-    }
 };
 
 }   // namespace NKikimr::NOlap::NReader::NSimple::NDuplicateFiltering
