+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/private_events.cpp
@@ -1,0 +1,3 @@
+#include "private_events.h"
+
+namespace NKikimr::NOlap::NReader::NSimple::NDuplicateFiltering::NPrivate {}
